Refactor contours correction in font processing
The contour correction process for OpenType-PS fonts was refactored. We moved the correction logic from `t2_charstrings.py` to `skia_tools.py` for better organization and moved the setting of `font.modified` to align with contour modification. This change should make the contour correction process clearer and more manageable. The `min_area_option` has been moved to `shared_options` module.
